+++ backend/app/main.py
@@ -1,13 +1,15 @@
-from fastapi import Depends, FastAPI, HTTPException, status
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from sqlalchemy.orm import Session
 
-from . import crud, models, schemas
-from .database import SessionLocal, engine
+from . import models
+from .database import engine
+from .routers import items, users
 
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
+app.include_router(users.router)
+app.include_router(items.router)
 
 origins = ["*"]
 
@@ -20,79 +22,6 @@ app.add_middleware(
 )
 
 
-# Dependency
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
 @app.get("/")
 def root() -> dict:
     return {"message": "Hello World"}
-
-
-@app.post(
-    "/users/", response_model=schemas.User, status_code=status.HTTP_201_CREATED
-)
-async def create_user(
-    user: schemas.UserCreate, db: Session = Depends(get_db)
-) -> schemas.User:
-    db_user = crud.get_user_by_email(db, email=user.email)
-    if db_user:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Email already registered",
-        )
-    return crud.create_user(db=db, user=user)
-
-
-@app.get(
-    "/users/",
-    response_model=list[schemas.User],
-    status_code=status.HTTP_200_OK,
-)
-async def read_users(
-    skip: int = 0, limit: int = 100, db: Session = Depends(get_db)
-) -> list[schemas.User]:
-    users = crud.get_users(db, skip=skip, limit=limit)
-    return users
-
-
-@app.get(
-    "/users/{user_id}",
-    response_model=schemas.User,
-    status_code=status.HTTP_200_OK,
-)
-async def read_user(
-    user_id: int, db: Session = Depends(get_db)
-) -> schemas.User:
-    db_user = crud.get_user(db, user_id=user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
-
-
-@app.post(
-    "/users/{user_id}/items/",
-    response_model=schemas.Item,
-    status_code=status.HTTP_201_CREATED,
-)
-async def create_item_for_user(
-    user_id: int, item: schemas.ItemCreate, db: Session = Depends(get_db)
-) -> schemas.Item:
-    return crud.create_user_item(db=db, item=item, user_id=user_id)
-
-
-@app.get(
-    "/items/",
-    response_model=list[schemas.Item],
-    status_code=status.HTTP_200_OK,
-)
-async def read_items(
-    skip: int = 0, limit: int = 100, db: Session = Depends(get_db)
-) -> list[schemas.Item]:
-    items = crud.get_items(db, skip=skip, limit=limit)
-    return items
