+++ backend/app/routers/users.py
@@ -1,0 +1,48 @@
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+
+from .. import crud, schemas
+from ..dependencies import get_db
+
+router = APIRouter(prefix="/users", tags=["users"])
+
+
+@router.post(
+    "/users/", response_model=schemas.User, status_code=status.HTTP_201_CREATED
+)
+async def create_user(
+    user: schemas.UserCreate, db: Session = Depends(get_db)
+) -> schemas.User:
+    db_user = crud.get_user_by_email(db, email=user.email)
+    if db_user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Email already registered",
+        )
+    return crud.create_user(db=db, user=user)
+
+
+@router.get(
+    "/users/",
+    response_model=list[schemas.User],
+    status_code=status.HTTP_200_OK,
+)
+async def read_users(
+    skip: int = 0, limit: int = 100, db: Session = Depends(get_db)
+) -> list[schemas.User]:
+    users = crud.get_users(db, skip=skip, limit=limit)
+    return users
+
+
+@router.get(
+    "/users/{user_id}",
+    response_model=schemas.User,
+    status_code=status.HTTP_200_OK,
+)
+async def read_user(
+    user_id: int, db: Session = Depends(get_db)
+) -> schemas.User:
+    db_user = crud.get_user(db, user_id=user_id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
+    return db_user
